requirements.txt
Django==5.0.10
djangorestframework==3.15.2
psycopg[binary]==3.2.3
dj-database-url==2.2.0
python-dotenv==1.0.1

pytest==8.3.3
pytest-django==4.9.0
pytest-cov==5.0.0
